remodeled database query and googlemaps api call
